feat(PostActions): use tooltip to denote copy action
